requirements.txt
python-telegram-bot==20.0a6
moviepy
